[ADD] - project_issue_deadline: Renamed module to project_issue_day_count
